+++ scrapy/spiders/sitemap.py
@@ -17,7 +17,7 @@ from typing import (
 
 from scrapy.http import Request, Response, XmlResponse
 from scrapy.spiders import Spider
-from scrapy.utils._compression import _DecompressionMaxSizeExceeded
+from scrapy.utils._compression_ops import _DecompressionMaxSizeExceeded
 from scrapy.utils.gz import gunzip, gzip_magic_number
 from scrapy.utils.sitemap import Sitemap, sitemap_urls_from_robots
