+++ scrapy/utils/_compression_ops.py
