+++ scrapy/utils/gz.py
@@ -5,7 +5,7 @@ from gzip import GzipFile
 from io import BytesIO
 from typing import TYPE_CHECKING
 
-from ._compression import _CHUNK_SIZE, _DecompressionMaxSizeExceeded
+from ._compression_ops import _CHUNK_SIZE, _DecompressionMaxSizeExceeded
 
 if TYPE_CHECKING:
     from scrapy.http import Response
